Implement modular UI structure with navigation and page management; add bulk download and hack history pages
