+++ config.json
@@ -1,5 +1,6 @@
 {
-  "base_rom_path": "",
-  "output_dir": "",
-  "api_delay": 0.8
-}
+  "base_rom_path": "C:/Users/georg/Desktop/SMWCentralDownloader/clean.smc",
+  "output_dir": "C:/Users/georg/Desktop/SMWCentralDownloader/Romhacks",
+  "api_delay": 0.8,
+  "root": {
+    "log_text":
