+++ ui/pages/__init__.py
@@ -1,0 +1,2 @@
+from .bulk_download_page import BulkDownloadPage
+from .hack_history_page import HackHistoryPage
