+++ ui/pages/hack_history_page.py
@@ -1,0 +1,27 @@
+from tkinter import ttk
+
+class HackHistoryPage:
+    """Hack history page implementation"""
+    
+    def __init__(self, parent):
+        self.parent = parent
+        self.frame = None
+    
+    def create(self):
+        """Create the hack history page"""
+        self.frame = ttk.Frame(self.parent, padding=25)
+        
+        # Add placeholder content
+        ttk.Label(
+            self.frame,
+            text="Hack History Page",
+            font=("Segoe UI", 14)
+        ).pack(pady=50)
+        
+        ttk.Label(
+            self.frame,
+            text="This page will be implemented in a future update.",
+            font=("Segoe UI", 10)
+        ).pack()
+        
+        return self.frame
